% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_matrix_function.R
\name{cor_matrix_function}
\alias{cor_matrix_function}
\title{NBA oldest_player_function Created by Hanzhang}
\usage{
cor_matrix_function(year)
}
\arguments{
\item{correlation_matrix}{Do you want to know the correlation matrix for the numeric parameters in the NBA dataset?}
}
\description{
This function allows you to create correlation matrix for the numeric parameters in the NBA dataset
}
\examples{
cor_matrix_function()
}
\keyword{NBA,}
\keyword{correlation}
\keyword{matrix}
\keyword{numeric,}
