% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/most_games_player_function.R
\name{most_games_player_function}
\alias{most_games_player_function}
\title{NBA most_games_player_function Created by Hanzhang}
\usage{
most_games_player_function(year)
}
\arguments{
\item{games}{Do you want to know who played most games in the NBA data for a chosen input year?}
}
\description{
This function allows you to know the player who played most games in the NBA data for a chosen input year
}
\examples{
most_games_player_function()
}
\keyword{Games}
\keyword{NBA,}
